requirements.txt
trimesh>=4.0.0
pyrender>=0.1.45
numpy>=1.24.0
Pillow>=10.0
scipy>=1.11.0
PyYAML>=6.0
pyglet<2.0
tqdm>=4.65
